CoursesWebApplication/Analytics.Service/requirements.txt
kafka-python
sqlalchemy
fastapi            
uvicorn            
